refactor: move multiple user errors to `src`
