+++ api/lib/domain/usecases/find-tutorials.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
 import { Scorecard } from '../../../src/evaluation/domain/models/Scorecard.js';
-import { UserNotAuthorizedToAccessEntityError } from '../errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../src/shared/domain/errors.js';
 import { KnowledgeElement } from '../models/KnowledgeElement.js';
 
 const findTutorials = async function ({
