+++ api/lib/domain/usecases/get-next-challenge-for-competence-evaluation.js
@@ -1,5 +1,6 @@
 import * as algorithmDataFetcherService from '../../../src/evaluation/domain/services/algorithm-methods/data-fetcher.js';
-import { AssessmentEndedError, UserNotAuthorizedToAccessEntityError } from '../errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../src/shared/domain/errors.js';
+import { AssessmentEndedError } from '../errors.js';
 
 const getNextChallengeForCompetenceEvaluation = async function ({
   pickChallengeService,
