+++ api/src/prescription/campaign-participation/domain/usecases/begin-campaign-participation-improvement.js
@@ -1,7 +1,5 @@
-import {
-  AlreadySharedCampaignParticipationError,
-  UserNotAuthorizedToAccessEntityError,
-} from '../../../../../lib/domain/errors.js';
+import { AlreadySharedCampaignParticipationError } from '../../../../../lib/domain/errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../shared/domain/errors.js';
 import { Assessment } from '../../../../shared/domain/models/Assessment.js';
 
 const beginCampaignParticipationImprovement = async function ({
