+++ api/src/prescription/campaign-participation/domain/usecases/compute-campaign-participation-analysis.js
@@ -1,5 +1,5 @@
-import { UserNotAuthorizedToAccessEntityError } from '../../../../../lib/domain/errors.js';
 import { CampaignLearningContent } from '../../../../../lib/domain/models/CampaignLearningContent.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../shared/domain/errors.js';
 import { CampaignParticipationDeletedError } from '../errors.js';
 
 const computeCampaignParticipationAnalysis = async function ({
