+++ api/src/prescription/campaign-participation/domain/usecases/get-campaign-profile.js
@@ -1,4 +1,4 @@
-import { UserNotAuthorizedToAccessEntityError } from '../../../../../lib/domain/errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../shared/domain/errors.js';
 
 const getCampaignProfile = async function ({
   userId,
