+++ api/src/prescription/campaign/domain/usecases/find-paginated-campaign-participants-activities.js
@@ -1,4 +1,4 @@
-import { UserNotAuthorizedToAccessEntityError } from '../../../../../lib/domain/errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../src/shared/domain/errors.js';
 
 const findPaginatedCampaignParticipantsActivities = async function ({
   userId,
