+++ api/src/prescription/campaign/domain/usecases/statistics/get-campaign-participations-activity-by-day.js
@@ -1,4 +1,4 @@
-import { UserNotAuthorizedToAccessEntityError } from '../../../../../../lib/domain/errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../../src/shared/domain/errors.js';
 
 const getCampaignParticipationsActivityByDay = async function ({
   userId,
