+++ api/src/prescription/campaign/domain/usecases/statistics/get-campaign-participations-counts-by-stage.js
@@ -1,4 +1,5 @@
-import { NoStagesForCampaign, UserNotAuthorizedToAccessEntityError } from '../../../../../../lib/domain/errors.js';
+import { NoStagesForCampaign } from '../../../../../../lib/domain/errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../../src/shared/domain/errors.js';
 
 const getCampaignParticipationsCountByStage = async function ({
   userId,
