+++ api/src/prescription/campaign/domain/usecases/statistics/get-participations-count-by-mastery-rate.js
@@ -1,4 +1,4 @@
-import { UserNotAuthorizedToAccessEntityError } from '../../../../../../lib/domain/errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../../src/shared/domain/errors.js';
 
 const getParticipationsCountByMasteryRate = async function ({
   campaignId,
