+++ api/src/prescription/organization-learner/domain/usecases/find-association-between-user-and-organization-learner.js
@@ -1,8 +1,5 @@
-import {
-  CampaignCodeError,
-  OrganizationLearnerDisabledError,
-  UserNotAuthorizedToAccessEntityError,
-} from '../../../../../lib/domain/errors.js';
+import { CampaignCodeError, OrganizationLearnerDisabledError } from '../../../../../lib/domain/errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../src/shared/domain/errors.js';
 
 const findAssociationBetweenUserAndOrganizationLearner = async function ({
   authenticatedUserId,
