+++ api/tests/prescription/campaign-participation/unit/domain/usecases/get-campaign-assessment-participation-result_test.js
@@ -1,5 +1,5 @@
-import { UserNotAuthorizedToAccessEntityError } from '../../../../../../lib/domain/errors.js';
 import { getCampaignAssessmentParticipationResult } from '../../../../../../src/prescription/campaign-participation/domain/usecases/get-campaign-assessment-participation-result.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | get-campaign-assessment-participation-result', function () {
