+++ api/tests/prescription/campaign-participation/unit/domain/usecases/get-campaign-profile_test.js
@@ -1,6 +1,6 @@
-import { UserNotAuthorizedToAccessEntityError } from '../../../../../../lib/domain/errors.js';
 import { getCampaignProfile } from '../../../../../../src/prescription/campaign-participation/domain/usecases/get-campaign-profile.js';
 import { LOCALE } from '../../../../../../src/shared/domain/constants.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 const { FRENCH_SPOKEN } = LOCALE;
