+++ api/tests/prescription/campaign/integration/domain/usecases/statistics/get-campaign-participations-counts-by-stage_test.js
@@ -1,5 +1,6 @@
-import { NoStagesForCampaign, UserNotAuthorizedToAccessEntityError } from '../../../../../../../lib/domain/errors.js';
+import { NoStagesForCampaign } from '../../../../../../../lib/domain/errors.js';
 import { usecases } from '../../../../../../../src/prescription/campaign/domain/usecases/index.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../../../src/shared/domain/errors.js';
 import {
   catchErr,
   databaseBuilder,
