+++ api/tests/prescription/campaign/integration/domain/usecases/statistics/get-campaign-participations-counts-by-status_test.js
@@ -1,6 +1,6 @@
-import { UserNotAuthorizedToAccessEntityError } from '../../../../../../../lib/domain/errors.js';
 import { usecases } from '../../../../../../../src/prescription/campaign/domain/usecases/index.js';
 import { CampaignParticipationStatuses } from '../../../../../../../src/prescription/shared/domain/constants.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../../../../src/shared/domain/errors.js';
 import { catchErr, databaseBuilder, expect } from '../../../../../../test-helper.js';
 
 const { STARTED, TO_SHARE } = CampaignParticipationStatuses;
