+++ api/tests/unit/application/error-manager_test.js
@@ -21,7 +21,6 @@ import {
   SendingEmailToInvalidDomainError,
   SendingEmailToInvalidEmailAddressError,
   UnexpectedUserAccountError,
-  UserHasAlreadyLeftSCO,
   UserShouldNotBeReconciledOnAnotherAccountError,
 } from '../../../lib/domain/errors.js';
 import { UnableToAttachChildOrganizationToParentOrganizationError } from '../../../src/organizational-entities/domain/errors.js';
@@ -108,19 +107,6 @@ describe('Unit | Application | ErrorManager', function () {
 
       // then
       expect(HttpErrors.UnauthorizedError).to.have.been.calledWithExactly(error.message);
-    });
-
-    it('should instantiate UnauthorizedError when UserHasAlreadyLeftSCO', async function () {
-      // given
-      const error = new UserHasAlreadyLeftSCO();
-      sinon.stub(HttpErrors, 'ForbiddenError');
-      const params = { request: {}, h: hFake, error };
-
-      // when
-      await handle(params.request, params.h, params.error);
-
-      // then
-      expect(HttpErrors.ForbiddenError).to.have.been.calledWithExactly(error.message);
     });
 
     it('should instantiate PreconditionFailedError when NotEnoughDaysPassedBeforeResetCampaignParticipationError', async function () {
