+++ api/tests/unit/domain/usecases/find-competence-evaluations-by-assessment_test.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
-import { UserNotAuthorizedToAccessEntityError } from '../../../../lib/domain/errors.js';
 import { findCompetenceEvaluationsByAssessment } from '../../../../lib/domain/usecases/find-competence-evaluations-by-assessment.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../src/shared/domain/errors.js';
 import { catchErr, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | UseCase | find-competence-evaluations-by-assessment', function () {
