+++ api/tests/unit/domain/usecases/find-tutorials_test.js
@@ -1,7 +1,7 @@
-import { UserNotAuthorizedToAccessEntityError } from '../../../../lib/domain/errors.js';
 import { KnowledgeElement } from '../../../../lib/domain/models/KnowledgeElement.js';
 import { findTutorials } from '../../../../lib/domain/usecases/find-tutorials.js';
 import { Scorecard } from '../../../../src/evaluation/domain/models/Scorecard.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../src/shared/domain/errors.js';
 import { domainBuilder, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | UseCase | find-tutorials', function () {
